Fix silent georeferencing corruption and harden overflow/bounds handling
read_geotiff.c: TIFFGetField(..., GTIFF_PIXELSCALE, ...) was called with a
fixed double[3] array instead of a double* out-pointer. Since TIFFGetField
is variadic and untyped, this compiled cleanly but silently read/wrote the
wrong memory, producing a consistent, wrong dx/dy for any GeoTIFF with a
valid pixel-scale tag. Traced this to a real-world incident: a geogrid
dataset built with this tool had its coordinate spacing off by ~1-2%
per pixel, shifting the coastline of a WRF domain built from it. Fixed by
declaring pixelscale as a pointer and checking the return value.

Also fixed the same silent-garbage-on-failure pattern for GTIFKeyGet()
(projection parameters and origin lat/lon), which wasn't the cause of the
above but is the identical failure mode, not yet triggered in practice.

Additional hardening from a follow-up audit:
- Removed the _CONV_BUF/_CONV_FILE macros in geogrid_tiles.c along with the
  dead get_tile_from_{d,i}/convert_from_{d,i} variants they generated (only
  the float path was ever called); widened the surviving pixel-count
  arithmetic to long/size_t to avoid 32-bit overflow on large rasters.
- write_geogrid.c: widened narray to long, check malloc() for NULL, and
  validate tile-boundary indices fit the fixed 5-digit filename format
  before formatting it (was an unchecked sprintf into a 24-byte buffer).
- convert_geotiff.c: length-check the filename/-u/-d arguments before
  copying into fixed 256-byte buffers; upper-bound the -t tile size to
  match the 5-digit filename format.
- read_geotiff.c: warn when the diagonal-step dx/dy fallback (only used
  when no pixel-scale tag is present) is taken, since it silently assumes
  no rotation/shear in the transform.

Verified each change against the exact GeoTIFF that produced the original
corrupted dataset: dx/dy now matches the source raster's true pixel size,
and each incremental hardening fix produces byte-identical tile output to
the previous verified build.

+++ geogrid_tiles.h
@@ -22,9 +22,9 @@ extern "C" {
   int nxtiles(const GeogridIndex);
   int nytiles(const GeogridIndex);
   int nzsize(const GeogridIndex);
-  int gettilestart(int,int,const GeogridIndex);
-  int globalystride(const GeogridIndex);
-  int globalzstride(const GeogridIndex);
+  long gettilestart(int,int,const GeogridIndex);
+  long globalystride(const GeogridIndex);
+  long globalzstride(const GeogridIndex);
   float *alloc_tile_buffer(const GeogridIndex);
   void get_tile_from_f(int,int,const GeogridIndex,const float*,float*);
   void convert_from_f(const GeogridIndex,const float*);
